Move report helper to report module and fixes
