+++ pepr/report.py
@@ -587,3 +587,90 @@ def report_generator(save_path, attack_subsections, pdf=False):
         doc.generate_pdf(save_path + "/attack_report", clean_tex=False)
     else:
         doc.generate_tex(save_path + "/attack_report")
+
+
+def plot_class_dist_histogram(attack_alias, class_data, save_path):
+    """
+    Plot a class distribution histogram.
+
+    Parameters
+    ----------
+    attack_alias : str
+        Alias for a specific instantiation of the class.
+    class_data : list
+        List of data to plot per class.
+    save_path : str
+        Path to save the plotted figure.
+
+    Returns
+    -------
+    str
+        Path to the saved figure.
+    """
+    import matplotlib.pyplot as plt
+
+    fig = plt.figure()
+    ax = plt.axes()
+    ax.hist(class_data, edgecolor="black")
+    ax.set_xlabel("Accuracy")
+    ax.set_ylabel("Number of Classes")
+    ax.set_axisbelow(True)
+    alias_no_spaces = str.replace(attack_alias, " ", "_")
+    path = f"fig/{alias_no_spaces}-hist.pdf"
+    fig.savefig(save_path + f"/{path}")
+    plt.close(fig)
+    return path
+
+
+def create_attack_pars_table(report_section, values, pars_descriptors):
+    """
+    Generate LaTex table from pars_descriptors with fancy parameter descriptions.
+
+    Parameters
+    ----------
+    report_section : pepr.report.ReportSection
+        ReportSection container object for the current attack.
+    values : dict
+        Dictionary with the values stored for the keys in `pars_descriptors`. If
+        `pars_descriptors` has an entry "max_iter", then `values` should also have an
+        entry "max_iter" but holding the argument value instead of a description.
+    pars_descriptors : dict
+        Dictionary of attack parameters and their description shown in the attack
+        report.
+
+    Returns
+    -------
+    int
+        Number of table rows added.
+    """
+    from pylatex import Command, Tabular
+
+    new_rows = len(pars_descriptors)
+
+    if new_rows > 0:
+        # Create table for the attack parameters.
+        new_rows = len(pars_descriptors)
+        with report_section.create(Tabular("|l|c|")) as tab_ap:
+            for key in pars_descriptors:
+                # Add table row per pars_descriptor entry
+                desc = pars_descriptors[key]
+                if key in values:
+                    if isinstance(values[key], float):
+                        value = str(round(values[key], 3))
+                    else:
+                        value = str(values[key])
+                    tab_ap.add_hline()
+                    tab_ap.add_row([desc, value])
+            tab_ap.add_hline()
+        report_section.append(Command("captionsetup", "labelformat=empty"))
+        report_section.append(
+            Command(
+                "captionof",
+                "table",
+                extra_arguments="Attack parameters",
+            )
+        )
+    else:
+        report_section.append("Attack has no configuration parameters.")
+
+    return new_rows
